Split policy::Config into its own module
